Move signing function to utils (mock in tests)
